+++ app/containers/AgentDesktop/reducer.js
@@ -104,7 +104,7 @@ import {
   RESET_FORM,
 } from './constants';
 
-// import { outboundConnectingVoiceInteraction, voiceInteraction, voiceInteractionWithTransfersAndScripts, emailInteraction, smsInteractionWithLotsOfMessagesAndScript, smsInteractionWithLotsOfMessagesAndScript2, scriptOnly } from './assets/mockInteractions'; // eslint-disable-line no-unused-vars
+// import { outboundConnectingVoiceInteraction, voiceInteraction, voiceInteractionWithTransfersAndScripts, emailInteraction, smsInteractionWithLotsOfMessagesAndScript, smsInteractionWithLotsOfMessagesAndScript2, smsInteractionWithLotsOfMessagesAndScript3, smsInteractionWithLotsOfMessagesAndScript4, smsInteractionWithLotsOfMessagesAndScript5, smsInteractionWithLotsOfMessagesAndScript6, scriptOnly } from './assets/mockInteractions'; // eslint-disable-line no-unused-vars
 
 const blankNewInteractionPanel = {
   interactionId: 'creating-new-interaction',
@@ -131,6 +131,10 @@ const initialState = fromJS({
     // emailInteraction,
     // smsInteractionWithLotsOfMessagesAndScript,
     // smsInteractionWithLotsOfMessagesAndScript2,
+    // smsInteractionWithLotsOfMessagesAndScript3,
+    // smsInteractionWithLotsOfMessagesAndScript4,
+    // smsInteractionWithLotsOfMessagesAndScript5,
+    // smsInteractionWithLotsOfMessagesAndScript6,
     // scriptOnly,
   ],
   noInteractionContactPanel: {